Rename termsenum -> terms_enum and move to 'functionality' directory
