+++ quesma/quesma/functionality/terms_enum/request.go
@@ -1,4 +1,4 @@
-package termsenum
+package terms_enum
 
 import (
 	"bytes"
