+++ quesma/quesma/functionality/terms_enum/terms_enum_test.go
@@ -1,4 +1,4 @@
-package termsenum
+package terms_enum
 
 import (
 	"context"
